+++ cmd/gizzi-code/packaging/chocolatey/tools/chocolateyinstall.ps1
@@ -3,8 +3,8 @@ $ErrorActionPreference = 'Stop'
 $packageName = 'gizzi-code'
 $toolsDir = "$(Split-Path -parent $MyInvocation.MyCommand.Definition)"
 # Release assets are version-named zips: gizzi-code-v<version>-windows-x64.zip
-$url64 = 'https://github.com/Gizziio/allternit-platform/releases/download/gizzi-code/v2.0.1/gizzi-code-v2.0.1-windows-x64.zip'
-$checksum64 = 'f4e0184f3a2ac046732b479874f6b35d8a24b368900680c2a781e30f276412be'
+$url64 = 'https://github.com/Gizziio/allternit-platform/releases/download/gizzi-code/v2.0.2/gizzi-code-v2.0.2-windows-x64.zip'
+$checksum64 = '3ea78cfe177c6c6cc3ebe5bed82dfc981a1087856d54cb2ba7956d3771e04423'
 
 $packageArgs = @{
   packageName    = $packageName